feature: mongoose refactor code

+++ src/modules/user/UserController.ts
@@ -15,35 +15,18 @@ class UserController {
         return UserController.instance;
     }
 
-    // eslint-disable-next-line class-methods-use-this
-    public list = async (req, res): Promise<IUser[]> => {
-        const { locals: { logger }, services } = res;
-        const { moduleService } = services;
-        try {
-            const { limit, skip } = req.query;
-            const { searchOptions } = req.body;
-            const result = await moduleService.list(
-                limit,
-                skip,
-                searchOptions,
-            );
-            if (!result.length) {
-                logger.debug({ message: 'Users not found', data: [] });
-                return res.send(SystemResponse.notFoundError('Users not found', []));
-            }
-            logger.info({ message: 'List of users', data: result });
-            return res.send(SystemResponse.success('List of users', result));
-        } catch (err) {
-            logger.error({ message: err.message, option: [{ Error: err.stack }] });
-            return res.send(SystemResponse.internalServerError(err._message, err.errors));
-        }
-    };
-
-    // eslint-disable-next-line class-methods-use-this
     public create = async (req, res) => {
         const { locals: { logger }, services } = res;
         const { moduleService } = services;
         try {
+            const { email } = req.body;
+            const totalUsers = await moduleService.count({ email });
+            if (totalUsers) {
+                logger.error({ message: 'User already exists' });
+                return res.send(
+                    SystemResponse.badRequestError('User already exists', {}),
+                );
+            }
             const result = await moduleService.create(
                 req.body,
             );
@@ -51,9 +34,9 @@ class UserController {
             return res.send(
                 SystemResponse.success('User added successfully', result),
             );
-        } catch (err) {
-            logger.error({ message: err.message, option: [{ Error: err.stack }] });
-            return res.send(SystemResponse.internalServerError(err._message, err.errors));
+        } catch (error) {
+            logger.error({ message: error.message, option: [{ Error: error.stack }] });
+            return res.send(SystemResponse.internalServerError(error.message, error));
         }
     };
 
@@ -61,15 +44,53 @@ class UserController {
         const { locals: { logger }, services } = res;
         const { moduleService } = services;
         try {
-            const { users } = req.body;
+            const { users = [] } = req.body;
+            const fetchAllEmails = users.map(({ email }) => email);
+            if (fetchAllEmails.length !== [...new Set(fetchAllEmails)].length) {
+                logger.error({ message: 'Duplicate User' });
+                return res.send(
+                    SystemResponse.badRequestError('Duplicate User', {}),
+                );
+            }
+            const totalUsers = await moduleService.count(
+                { email: { $in: [...new Set(fetchAllEmails)] } },
+            );
+            if (totalUsers) {
+                logger.error({ message: 'User already exists' });
+                return res.send(
+                    SystemResponse.badRequestError('User already exists', {}),
+                );
+            }
             const result = await moduleService.bulkInsert(users);
             logger.info({ message: 'Users added successfully', data: result });
             return res.send(
                 SystemResponse.success('Users added successfully', result),
             );
-        } catch (err) {
-            logger.error({ message: err.message, option: [{ Error: err.stack }] });
-            return res.send(SystemResponse.internalServerError(err._message, err.errors));
+        } catch (error) {
+            logger.error({ message: error.message, option: [{ Error: error.stack }] });
+            return res.send(SystemResponse.internalServerError(error.message, error));
+        }
+    };
+
+    public list = async (req, res): Promise<IUser[]> => {
+        const { locals: { logger }, services } = res;
+        const { moduleService } = services;
+        try {
+            const { limit = 10, skip = 0, name } = req.query;
+            const result = await moduleService.list(
+                limit,
+                skip,
+                name && { name },
+            );
+            if (!result.length) {
+                logger.debug({ message: 'Users not found', data: [] });
+                return res.send(SystemResponse.notFoundError('Users not found', []));
+            }
+            logger.info({ message: 'List of users', data: result });
+            return res.send(SystemResponse.success('List of users', result));
+        } catch (error) {
+            logger.error({ message: error.message, option: [{ Error: error.stack }] });
+            return res.send(SystemResponse.internalServerError(error.message, error));
         }
     };
 
@@ -87,9 +108,9 @@ class UserController {
             }
             logger.info({ message: 'Fetch user successfully', data: userInfo });
             return res.send(SystemResponse.success('Fetch user successfully', userInfo));
-        } catch (err) {
-            logger.error({ message: err._message, option: [{ Error: err.stack }] });
-            return res.send(SystemResponse.internalServerError(err._message, err.errors));
+        } catch (error) {
+            logger.error({ message: error.message, option: [{ Error: error.stack }] });
+            return res.send(SystemResponse.internalServerError(error.message, error));
         }
     };
 
@@ -99,6 +120,16 @@ class UserController {
         try {
             const { id } = req.params;
             const data = req.body;
+            if (data?.email) {
+                const { email } = data;
+                const totalUsers = await moduleService.count({ email });
+                if (totalUsers) {
+                    logger.error({ message: 'User already exists' });
+                    return res.send(
+                        SystemResponse.badRequestError('User already exists', {}),
+                    );
+                }
+            }
             const user = await moduleService.count({
                 id,
             });
@@ -106,14 +137,14 @@ class UserController {
                 logger.info({ message: 'User not found', data: {} });
                 return res.send(SystemResponse.notFoundError('User not found', {}));
             }
-            const result = await moduleService.update(req.params.id, data, { deletedAt: 0 });
+            const result = await moduleService.update(req.params.id, data);
             logger.info({ messgae: 'Userinfo update successfully', data: result });
             return res.send(
                 SystemResponse.success('Userinfo update successfully', result),
             );
-        } catch (err) {
-            logger.error({ message: err.message, option: [{ Error: err.stack }] });
-            return res.send(SystemResponse.internalServerError(err._message, err.errors));
+        } catch (error) {
+            logger.error({ message: error.message, option: [{ Error: error.stack }] });
+            return res.send(SystemResponse.internalServerError(error.message, error));
         }
     };
 
@@ -122,6 +153,11 @@ class UserController {
         const { moduleService } = services;
         try {
             const { name } = req.query;
+            const data = req.body;
+            if (data?.email) {
+                logger.info({ message: 'Cannot update duplicate email', data: [] });
+                return res.send(SystemResponse.notFoundError('Cannot update duplicate emails', []));
+            }
             const totalUsers = await moduleService.count({ name });
             if (!totalUsers) {
                 logger.info({ message: 'User not found', data: [] });
@@ -129,11 +165,11 @@ class UserController {
             }
             const result = await moduleService.bulkUpdate(
                 { name },
-                { $inc: { age: 10 } },
+                data,
             );
-            logger.info({ messgae: 'Userinfo update successfully', data: result });
+            logger.info({ messgae: 'Usersinfo update successfully', data: result });
             return res.send(
-                SystemResponse.success('Userinfo update successfully', result),
+                SystemResponse.success('Usersinfo update successfully', result),
             );
         } catch (error) {
             logger.error({ message: error.message, option: [{ Error: error.stack }] });
@@ -146,14 +182,19 @@ class UserController {
         const { moduleService } = services;
         try {
             const { id } = req.params;
+            const totalUser = await moduleService.count({ id });
+            if (!totalUser) {
+                logger.info({ message: 'User not found', data: [] });
+                return res.send(SystemResponse.notFoundError('User not found', []));
+            }
             const result = await moduleService.delete({
                 id,
             });
             logger.info({ message: 'User information deleted' });
             return res.send(SystemResponse.success('User information deleted', result));
-        } catch (err) {
-            logger.error({ message: err.message, option: [{ Error: err.stack }] });
-            return res.send(SystemResponse.internalServerError(err._message, err.errors));
+        } catch (error) {
+            logger.error({ message: error.message, option: [{ Error: error.stack }] });
+            return res.send(SystemResponse.internalServerError(error.message, error));
         }
     };
 
@@ -162,12 +203,17 @@ class UserController {
         const { moduleService } = services;
         try {
             const { name } = req.query;
-            const result = await moduleService.bulkDelete(name);
-            logger.info({ message: 'User information deleted' });
-            return res.send(SystemResponse.success('User information deleted', result));
-        } catch (err) {
-            logger.error({ message: err.message, option: [{ Error: err.stack }] });
-            return res.send(SystemResponse.internalServerError(err._message, err.errors));
+            const totalUsers = await moduleService.count({ name });
+            if (!totalUsers) {
+                logger.info({ message: 'Users not found', data: [] });
+                return res.send(SystemResponse.notFoundError('Users not found', []));
+            }
+            const result = await moduleService.bulkDelete({ name });
+            logger.info({ message: 'Users information deleted' });
+            return res.send(SystemResponse.success('Users information deleted', result));
+        } catch (error) {
+            logger.error({ message: error.message, option: [{ Error: error.stack }] });
+            return res.send(SystemResponse.internalServerError(error.message, error));
         }
     };
 }
